Create 'elements' folder in 'component' folder and move related files into it.
