+++ frontend/src/app/components/ClassDetail.tsx
@@ -10,7 +10,7 @@ import {
   formatTime24Hour,
 } from "../helper/dateUtils";
 import Image from "next/image";
-import ActionButton from "./ActionButton";
+import ActionButton from "./elements/buttons/actionButton/ActionButton";
 import styles from "./ClassDetail.module.scss";
 import {
   CalendarDaysIcon,
