+++ frontend/src/app/components/ClassesTable.tsx
@@ -4,7 +4,7 @@ import {
   formatShortDate,
   formatTime24Hour,
 } from "../helper/dateUtils";
-import ActionButton from "./ActionButton";
+import ActionButton from "./elements/buttons/actionButton/ActionButton";
 import styles from "./ClassesTable.module.scss";
 import Image from "next/image";
 import { InformationCircleIcon } from "@heroicons/react/24/solid";
