+++ frontend/src/app/components/admins-dashboard/InstructorSearch.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import styles from "./InstructorSearch.module.scss";
 import { getInstructors } from "@/app/helper/instructorsApi";
-import ActionButton from "@/app/components/ActionButton";
+import ActionButton from "@/app/components/elements/buttons/actionButton/ActionButton";
 
 function InstructorSearch({
   handleSendInstructor,
