+++ frontend/src/app/components/customers-dashboard/regular-classes/RegularClasses.tsx
@@ -4,10 +4,10 @@ import { useEffect, useState } from "react";
 import styles from "./RegularClasses.module.scss";
 import { PlusIcon } from "@heroicons/react/24/outline";
 import { getSubscriptionsByCustomerId } from "@/app/helper/subscriptionsApi";
-import ActionButton from "@/app/components/ActionButton";
+import ActionButton from "@/app/components/elements/buttons/actionButton/ActionButton";
 import CurrentSubscription from "@/app/components/customers-dashboard/regular-classes/CurrentSubscription";
 import AddSubscription from "@/app/components/customers-dashboard/regular-classes/AddSubscription";
-import Loading from "../../Loading";
+import Loading from "../../elements/loading/Loading";
 
 function RegularClasses({
   customerId,
