+++ frontend/src/app/components/elements/buttons/redirectButton/RedirectButton.module.scss
@@ -1,0 +1,175 @@
+@import "../../../../variables.module.scss";
+
+.btnComponent {
+  display: inline-block;
+  align-items: center;
+  border-radius: 0.375rem;
+  padding: 0.5rem 1rem;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: #ffffff;
+  text-decoration: none;
+  height: fit-content;
+
+  &:hover {
+    cursor: pointer;
+  }
+
+  &.disabled {
+    color: gray;
+    border-color: #e0e0e0;
+    cursor: not-allowed;
+    pointer-events: none;
+    background-color: #f3f4f6;
+
+    &:hover {
+      background-color: #f3f4f6;
+      border-color: gray;
+    }
+  }
+}
+
+.content {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.icon {
+  height: 1.25rem;
+  margin-left: 0.5rem;
+}
+
+.backBtn,
+.cancelBtn,
+.deleteBtn {
+  background-color: #f3f4f6;
+  transition:
+    background-color 0.3s ease,
+    transform 0.3s ease;
+  border: 1px solid #6b7280;
+  color: #6b7280;
+  padding: 0.5rem 1rem;
+  border-radius: 0.375rem;
+
+  &:hover {
+    background-color: #e5e7eb;
+    transform: scale(1.02);
+  }
+
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 3px rgba(176, 176, 176, 0.5);
+  }
+}
+
+.rescheduleBtn,
+.editBtn,
+.saveBtn {
+  background-color: #28a745;
+  transition:
+    background-color 0.3s ease,
+    transform 0.3s ease;
+  border: 1px solid transparent;
+  color: #ffffff;
+  padding: 0.5rem 1rem;
+  border-radius: 0.375rem;
+
+  &:hover {
+    background-color: #218838;
+    transform: scale(1.02);
+  }
+
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 3px rgba(40, 167, 69, 0.5);
+  }
+}
+
+.bookBtn,
+.addBtn {
+  background-color: $blue_primary;
+  transition:
+    background-color 0.3s ease,
+    transform 0.3s ease;
+  border: 1px solid transparent;
+  color: #ffffff;
+  padding: 0.5rem 1rem;
+  border-radius: 0.375rem;
+
+  &:hover {
+    background-color: $blue_primary_hover;
+  }
+
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 3px rgba(38, 143, 255, 0.5);
+  }
+}
+
+.back {
+  background-color: $back-btn-bc;
+  transition:
+    background-color 0.3s ease,
+    transform 0.3s ease;
+  border: 1px solid transparent;
+  color: #333333;
+  padding: 0.5rem 0;
+  border-radius: 10px;
+  width: 140px;
+
+  &:hover {
+    background-color: $back-btn-bc-hover;
+  }
+
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 3px rgba(38, 143, 255, 0.5);
+  }
+}
+
+.bookClass {
+  background-color: $green-success;
+  transition:
+    background-color 0.3s ease,
+    transform 0.3s ease;
+  border: 1px solid transparent;
+  color: white;
+  padding: 0.5rem 1rem;
+  border-radius: 100px;
+  box-shadow:
+    rgba(50, 50, 93, 0.25) 0px 2px 5px -1px,
+    rgba(0, 0, 0, 0.3) 0px 1px 3px -1px; // width: 140px;
+
+  &:hover {
+    background-color: $green-success-hover;
+  }
+
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 3px rgba(38, 143, 255, 0.5);
+  }
+}
+
+.deleteChild,
+.cancelEditingChild {
+  background-color: #d9d9d9;
+  transition:
+    background-color 0.3s ease,
+    transform 0.3s ease;
+  border: 1px solid transparent;
+  color: #333333;
+  padding: 0.5rem 0;
+  border-radius: 10px;
+  width: 130px;
+  height: 35px;
+
+  &:hover {
+    background-color: $back-btn-bc-hover;
+  }
+
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 3px rgba(38, 143, 255, 0.5);
+  }
+}
