+++ frontend/src/app/components/elements/modal/Modal.module.scss
