+++ frontend/src/app/customers/[id]/classes/[classId]/page.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState, useCallback } from "react";
 import { cancelClass, getClassesByCustomerId } from "@/app/helper/classesApi";
 import ClassDetail from "@/app/components/ClassDetail";
 import ClassesTable from "@/app/components/ClassesTable";
-import RedirectButton from "@/app/components/RedirectButton";
+import RedirectButton from "@/app/components/elements/buttons/redirectButton/RedirectButton";
 import {
   isPastClassDateTime,
   isPastPreviousDayDeadline,
