+++ frontend/src/app/customers/register/page.tsx
@@ -7,14 +7,14 @@ import { prefectures } from "@/app/helper/data";
 import styles from "./page.module.scss";
 import Image from "next/image";
 import Link from "next/link";
-import TextInput from "@/app/components/TextInput";
+import TextInput from "@/app/components/elements/textInput/TextInput";
 import {
   EnvelopeIcon,
   HomeIcon,
   LockClosedIcon,
   UserCircleIcon,
 } from "@heroicons/react/24/outline";
-import ActionButton from "@/app/components/ActionButton";
+import ActionButton from "@/app/components/elements/buttons/actionButton/ActionButton";
 import { toast, ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import { useSelect } from "@/app/hooks/useSelect";
