+++ frontend/src/app/instructors/login/page.tsx
@@ -6,9 +6,9 @@ import { useInput } from "@/app/hooks/useInput";
 import { isValidLogin } from "@/app/helper/validationUtils";
 import styles from "./page.module.scss";
 import Image from "next/image";
-import TextInput from "@/app/components/TextInput";
+import TextInput from "@/app/components/elements/textInput/TextInput";
 import { EnvelopeIcon, LockClosedIcon } from "@heroicons/react/24/outline";
-import ActionButton from "@/app/components/ActionButton";
+import ActionButton from "@/app/components/elements/buttons/actionButton/ActionButton";
 import { toast, ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
